requirements.txt
Django==5.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0
python-decouple==3.8
Pillow==10.1.0
beautifulsoup4==4.13.4
matplotlib==3.10.3
pandas==2.3.0
plotly==6.1.2
seaborn==0.13.2
reportlab==4.4.2
fpdf2==2.8.3
openpyxl==3.1.5
requests==2.32.4
Markdown==3.8

